tipc: Yet more trivial compiler warning elimination
commit f817051fb94257220a54e702f9aad14fbdf9ab9c from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Gets rid of a pair of compiler warnings relating to possible
truncation of high order bits -- neither of which could actually
occur.

Integrated-by: Howard Xu <hao.xu@windriver.com>

+++ net/tipc/tipc_net.c
@@ -171,7 +171,7 @@ static void net_insert_element(struct net_element *e_ptr, net_element_set_t *set
 	for (i = 0; i < set->first_free; i++) {
 		dbg_assert(set->element[i]->addr != e_ptr->addr);
 		if (set->element[i]->addr > e_ptr->addr) {
-			int sz = (set->first_free - i) * sizeof(e_ptr);
+			size_t sz = (set->first_free - i) * sizeof(e_ptr);
 			memmove(&set->element[i+1], &set->element[i], sz);
 			break;
 		}
